requirements.txt
# Streamlit for web app
streamlit==1.30.0

# Data handling
pandas==2.1.0
numpy==1.26.5

# Plotting
plotly==5.21.0

# Environment variables
python-dotenv==1.1.1

# Machine Learning / NLP
scikit-learn==1.3.2
sentence-transformers==2.2.2
rank-bm25==0.2.2
cross-encoder==2.3.1

# ChromaDB vector store
chromadb==0.4.0

# Google Generative AI SDK
google-generativeai==0.8.5

# Logging (built-in, no need to install)
# os, json, math, pickle, time, random (built-in)

# Your local modules (no need to install)
# main, preprocess, vector_store, embedder, config, prompts
